[BE: A5] #S06P31A107-99 다이어리 저장하기 api 구현

Closes S06P31A107-99

See merge request s06-final/S06P31A107!24

+++ backend/src/main/java/com/ssafy/onda/global/common/entity/Image.java
@@ -26,7 +26,7 @@ public class Image extends BaseMemoEntity {
     private String savedImagePath;
 
     @Builder
-    public Image(Double x, Double y, Double width, Double height, Long imageSeq, String originImageName, String encodingImageName, String savedImagePath) {
+    public Image(Long x, Long y, Long width, Long height, Long imageSeq, String originImageName, String encodingImageName, String savedImagePath) {
         super(x, y, width, height);
         this.imageSeq = imageSeq;
         this.originImageName = originImageName;
